Moved Actions into data/types.py and renamed some enums
