+++ onyx/data/types.py
@@ -1,7 +1,24 @@
 from enum import Enum
 
 
-class OnyxScope(Enum):
+class Actions(Enum):
+    ACCESS = ("access", "access")
+    NO_ACCESS = ("no_access", "not access")
+    GET = ("get", "get")
+    LIST = ("list", "list")
+    FILTER = ("filter", "filter")
+    HISTORY = ("history", "get the history of")
+    IDENTIFY = ("identify", "identify values from")
+    ADD = ("add", "add")
+    CHANGE = ("change", "change")
+    DELETE = ("delete", "delete")
+
+    def __init__(self, label: str, description: str) -> None:
+        self.label = label
+        self.description = description
+
+
+class Scopes(Enum):
     ADMIN = "admin"
     UPLOADER = "uploader"
     ANALYST = "analyst"
@@ -10,7 +27,7 @@ class OnyxScope(Enum):
         self.label = label
 
 
-class OnyxObject(Enum):
+class Objects(Enum):
     RECORD = "records"
     ANALYSIS = "analyses"
